Move KASSERT test functionality to the crtp helper
